+++ init/init_db.py
@@ -1,7 +1,7 @@
 # Package: BulkDNS
 # Module: init/init_db
 # Author: Michal Selma <michal@selma.cc>
-# Rev: 2023-12-23
+# Rev: 2024-02-16
 
 import logging
 
@@ -22,7 +22,7 @@ def create_db(db, new_db_name):
         log.critical(f'Error: Incorrect database type. (Should not see me!)')
 
 
-def create_domain_tbl(db, tbl_names, tld):
+def create_tbl(db, tbl_names, tld):
     for tbl_name in tbl_names:
         table = f'{tbl_name}_{tld}'
         query = (f'CREATE TABLE {table} (domain VARCHAR(300) PRIMARY KEY NOT NULL, name VARCHAR(255), '
@@ -31,30 +31,10 @@ def create_domain_tbl(db, tbl_names, tld):
         db.execute_single(query, call_id)
 
 
-def initialize(db, tbl_names, tld):
+def initialize(db, new_db_name, tbl_names, tld):
     # Main operations database for char combinations, taken domains to be moved then to '_taken' db in separate process
-    new_db_name = 'domain'
     log.info(f'Creating database: {new_db_name}')
     create_db(db, new_db_name)
     db.db_name = new_db_name  # When new db has been created, for tables initialization modify db object to use new db
-    log.info(f'Creating tables structure...')
-    create_domain_tbl(db, tbl_names, tld)
-
-    # Below is to store domains that are taken. Archive and bring back if needed based on domain status change to
-    # be managed by archiver module.
-    new_db_name = 'domain_taken'
-    log.info(f'Creating database: {new_db_name}')
-    create_db(db, new_db_name)
-    db.db_name = new_db_name  # When new db is created, for tables initialization modify db object to use new db
-    log.info(f'Creating tables structure...')
-    create_domain_tbl(db, tbl_names, tld)
-
-    # Below is backup DB. Data backup is managed by archiver module.
-    new_db_name = 'domain_backup'
-    log.info(f'Creating database: {new_db_name}')
-    create_db(db, new_db_name)
-    db.db_name = new_db_name  # When new db is created, for tables initialization modify db object to use new db
-    log.info(f'Creating tables structure...')
-    create_domain_tbl(db, tbl_names, tld)
-
-    log.info(f'Finished')
+    log.info(f'Creating tables structure in {new_db_name}...')
+    create_tbl(db, tbl_names, tld)